requirements/dev-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/dev-requirements.txt requirements/dev-requirements.in
#
appdirs==1.4.4            # via black, virtualenv
astroid==2.4.2            # via pylint
attrs==19.3.0             # via black, flake8-eradicate, pytest
autoflake==1.3.1          # via -r requirements/dev-requirements.in
bandit==1.6.2             # via -r requirements/dev-requirements.in
black==19.10b0            # via -r requirements/dev-requirements.in
cfgv==3.2.0               # via pre-commit
click==7.1.2              # via black
colorama==0.4.3           # via radon
coverage==5.2.1           # via pytest-cov
distlib==0.3.1            # via virtualenv
eradicate==1.0            # via flake8-eradicate
filelock==3.0.12          # via virtualenv
flake8-blind-except==0.1.1  # via -r requirements/dev-requirements.in
flake8-breakpoint==1.1.0  # via -r requirements/dev-requirements.in
flake8-builtins==1.5.3    # via -r requirements/dev-requirements.in
flake8-comprehensions==3.2.3  # via -r requirements/dev-requirements.in
flake8-eradicate==0.4.0   # via -r requirements/dev-requirements.in
flake8-flask==0.9.3       # via -r requirements/dev-requirements.in
flake8-logging-format==0.6.0  # via -r requirements/dev-requirements.in
flake8-plugin-utils==1.3.1  # via flake8-breakpoint, flake8-pytest-style, flake8-return
flake8-polyfill==1.0.2    # via pep8-naming, radon
flake8-print==3.1.4       # via -r requirements/dev-requirements.in
flake8-pytest-style==1.2.3  # via -r requirements/dev-requirements.in
flake8-pytest==1.3        # via -r requirements/dev-requirements.in
flake8-return==1.1.2      # via -r requirements/dev-requirements.in
flake8==3.8.3             # via -r requirements/dev-requirements.in, flake8-builtins, flake8-comprehensions, flake8-eradicate, flake8-flask, flake8-polyfill, flake8-print, flake8-pytest
future==0.18.2            # via radon
gitdb==4.0.5              # via gitpython
gitpython==3.1.7          # via bandit
identify==1.4.25          # via pre-commit
iniconfig==1.0.1          # via pytest
isort==4.3.21             # via -r requirements/dev-requirements.in, pylint
lazy-object-proxy==1.4.3  # via astroid
mando==0.6.4              # via radon
mccabe==0.6.1             # via flake8, pylint
more-itertools==8.4.0     # via pytest
mypy-extensions==0.4.3    # via mypy
mypy==0.782               # via -r requirements/dev-requirements.in
nodeenv==1.4.0            # via pre-commit
packaging==20.4           # via pytest
pathspec==0.8.0           # via black
pbr==5.4.5                # via stevedore
pep8-naming==0.11.1       # via -r requirements/dev-requirements.in
pluggy==0.13.1            # via pytest
pre-commit==2.6.0         # via -r requirements/dev-requirements.in
py==1.9.0                 # via pytest
pycodestyle==2.6.0        # via flake8, flake8-print
pyflakes==2.2.0           # via autoflake, flake8
pylint==2.5.3             # via -r requirements/dev-requirements.in
pyparsing==2.4.7          # via packaging
pytest-cov==2.10.1        # via -r requirements/dev-requirements.in
pytest==6.0.1             # via -r requirements/dev-requirements.in, pytest-cov
pyyaml==5.3.1             # via bandit, pre-commit
r2c-py-ast==0.1.0b1       # via flake8-flask
radon==4.2.0              # via -r requirements/dev-requirements.in
regex==2020.7.14          # via black
six==1.15.0               # via astroid, bandit, flake8-print, mando, packaging, virtualenv
smmap==3.0.4              # via gitdb
stevedore==3.2.0          # via bandit
toml==0.10.1              # via black, pre-commit, pylint, pytest
typed-ast==1.4.1          # via black, mypy
typing-extensions==3.7.4.2  # via mypy
virtualenv==20.0.30       # via pre-commit
wrapt==1.12.1             # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
